+++ NEWS.md
@@ -14,6 +14,7 @@
 * new sorting option `"prop_lower"` for `gglikert()` (#62)
 * new helper `symmetric_limits()` to make a scale symmetric (#66)
 * new argument `symmetric` for `gglikert()` (#66)
+* new argument `default_by` for `stat_prop()` (#69)
 
 # ggstats 0.6.0